requirements.txt
pillow
flask
requests
pyyaml